CLN/TST: DRY skip_if_no_ne
